CognitiveServices/CustomVision/requirements.txt
azure-cognitiveservices-vision-customvision